Politicas de acceso
Proteger: eliminar

+++ app/Http/Controllers/RepositoryController.php
@@ -76,8 +76,12 @@ class RepositoryController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Repository $repository)
+    public function destroy(Request $request, Repository $repository)
     {
+        if ($request->user()->id != $repository->user_id) {
+            abort(403);
+        }
+
         $repository->delete();
 
         return redirect()->route('repositories.index');
